requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.1
pytest>=8.0.0
